# DriverLocation/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Driver Location API",
        "title": "Driver Location API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "paths": {
        "/api/driver-locations": {
            "post": {
                "description": "Creates driver location by longitude and latitude or bulk file.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Create Driver Location.",
                "parameters": [
                    {
                        "type": "number",
                        "description": "latitude",
                        "name": "latitude",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "longitude",
                        "name": "longitude",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "bulk location file",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/search/location": {
            "get": {
                "description": "gets nearest driver location by given point and radius",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Gets nearest driver location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "radius",
                        "name": "radius",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "longitude",
                        "name": "longitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "latitude",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.NearestDriverResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.NearestDriverResponse": {
            "type": "object",
            "properties": {
                "Title": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                }
            }
        }
    }
}